requirements.txt
requests
streamlit
fastapi
flask
tqdm
uvicorn
torch
seleniumbase
pyspark
flask-cors
python-logging-loki
python-dotenv
tensorboard
